application/pyproject.toml
[tool.poetry]
name = "addres_convert"
version = "1.0.1"
description = "pythonテンプレート"
authors = ["Taku Ikegami"]

[tool.poetry.dependencies]
python = "^3.10.8"
Django = "^4.1.3"
psycopg2 = "^2.9.5"
uWSGI = "^2.0.21"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.25.1"
boto3 = "^1.26.79"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pylint = "^2.15.5"
pytest-cov = "^4.0.0"
toml = "^0.10.2"
black = "^22.10.0"
isort = "^5.10.1"
pdoc3 = "^0.10.0"
mypy = "^0.991"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.py$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.mypy]
packages = ["application"]
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true


[logging]
version = 1
[logging.formatters.simple]
format = "[%(levelname)s] %(name)s %(asctime)s - %(message)s "
[logging.handlers.consoleHandler]
class = "logging.StreamHandler"
level = "DEBUG"
formatter = "simple"
stream = "ext://sys.stdout"
[logging.handlers.applicationHandler]
class = "logging.handlers.TimedRotatingFileHandler"
when = "D"
level = "INFO"
formatter = "simple"
filename = "./output/application.log"
[logging.handlers.errorHandler]
class = "logging.handlers.TimedRotatingFileHandler"
when = "D"
level = "ERROR"
formatter = "simple"
filename = "./output/emergency.log"
[logging.loggers.console]
level = "DEBUG"
handlers = ["consoleHandler"]
propagate = "no"
[logging.loggers.application]
level = "DEBUG"
handlers = ["consoleHandler","applicationHandler"]
propagate = "no"
[logging.loggers.emergency]
level = "DEBUG"
handlers = ["consoleHandler","errorHandler"]
propagate = "no"